Build for JVM and iOSX64
