+++ gradle.properties
@@ -1,4 +1,5 @@
 org.gradle.parallel=true
 org.gradle.caching=true
-kotlin.code.style = official
-libraryVersion = "0.0.1-SNAPSHOT"
+kotlin.code.style=official
+
+libraryVersion=0.0.1-SNAPSHOT
