+++ settings.gradle.kts
@@ -1,4 +1,4 @@
 rootProject.name = "bip21"
 
 plugins {}
-include("lib")
+// include("lib")
