+++ src/commonMain/kotlin/org/bitcointools/bip21/Exceptions.kt
